app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
requests==2.31.0
openai==0.28   # changed from 1.3.5 to 0.28 to support ChatCompletion
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whisper
torch
torchaudio
numpy
python-telegram-bot
